Properly clear out popup quests
This had a weird sideeffect of attempting to show the quest details
when opening the map because the map triggered QLU, which triggered
this logic showing the quest details, but the quest had no details since
it was already accepted, which hid the map, leaving no trace.
The initial QLU we were looking for triggered too early for the poups to
show up, so this never worked correctly anyways.

Fixes #45