Essentially rewrite the quest handling for TWW
This involves using the new settings system, and should properly deal
with caches using callbacks, something the old structure didn't allow.

It also uses new API introducted either in TWW or in earlier patches to
smooth out some edges, but there are still a few edge cases that needs
to be tested.