Expand quest automation options
It didn't make much sense to split accept and complete for
non-repeatable quests when repeatable quests were combined.
Also account for the fact that people never read changelogs, so ditch
the minimap warband tracking option as the primary way to toggle this
option, while leaving it in for the more advanced users, and also be a
bit more verbose about its existence in the options.

Also fixes a few minor issues with ignore list not being cleared
correctly when changing certain options.